fix: double timeout budget for reasoning model quality
Codex tool_timeout: 120 → 180 (set by glance install)
Transport guard: 115 → 175
Sub-agent deadline: 115 → 175
Chat timeout: 115 → 170
HTTP timeout: 60 → 90

v4-pro now gets 90s per HTTP call with room for 2-3 reasoning
iterations. The fallback to v4-flash only triggers after 2 consecutive
HTTP timeouts, preserving v4-pro quality for as long as possible.

+++ src/install/codex.rs
@@ -40,7 +40,7 @@ pub fn install() -> Result<()> {
     let mut glance = toml::value::Table::new();
     glance.insert("command".into(), toml::Value::String("glance-mcp".into()));
     glance.insert("startup_timeout_sec".into(), toml::Value::Integer(10));
-    glance.insert("tool_timeout_sec".into(), toml::Value::Integer(120));
+    glance.insert("tool_timeout_sec".into(), toml::Value::Integer(180));
     servers_tbl.insert("glance".into(), toml::Value::Table(glance));
 
     let serialized = toml::to_string_pretty(&doc).context("serialize codex config")?;
